pyproject.toml
[tool.poetry]
name = "gate-matrix-fall-detection"
version = "0.1.0"
description = "Fall detection using gate matrix"
authors = ["Vinnter Developers"]
classifiers=[
    "Programming Language :: Python :: 3.10",
]
readme = "README.md"
keywords = ["serverless"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
aws-lambda-powertools =  {extras = ["all"],version = "*"}
pandas = "2.1.4"
scipy = "^1.12.0"
matplotlib = "^3.8.2"
plotly-express = "^0.4.1"
numpy = "1.24.4"
seaborn = "^0.13.2"
pyqt6 = "^6.6.1"
loguru = "0.7.2"
bleak = "^0.21.1"
pyarrow = "^15.0.0"
boto3 = "1.33.13"
aws-cdk-core = "^1.204.0"
aws-cdk-aws-codecommit = "^1.204.0"
aws-cdk-pipelines = "^1.204.0"
aws-cdk-aws-s3 = "^1.204.0"
black = "24.1.1"
isort = "^5.13.2"
python-dotenv = "1.0.0"
moto = "4.2.13"
pytest = "7.4.4"
nbformat = "5.9.2"
plotly = "^5.18.0"
apache-airflow = "2.8.1"
flask-session = "0.5.0"
kubernetes = "^29.0.0"
apache-airflow-providers-amazon = "^8.17.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "3.6.0"
ipykernel = "6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
